+++ app/templates/request-code-enter-address.html
@@ -14,55 +14,11 @@
     }
 } %}
 
+{% set question_title = _('What is your postcode') %}
+{% set question_description = _('<p>To text you a new code we need to know the address for which you are answering.</p>') %}
+
 {% block main %}
 
-    {% if messages_dict %}
-        {% include 'partials/messages.html' with context %}
-    {% endif %}
-
-    {% call onsQuestion({
-        'title': _('What is your postcode?'),
-        'description': _('<p>To text you a new code we need to know the address for which you are answering.</p>')
-    }) %}
-
-        {% if 'postcode' in field_messages_dict %}
-            {{
-                onsInput({
-                    'id': 'postcode',
-                    'type': 'postcode',
-                    'autocomplete': 'postcode',
-                    'classes': 'input--w-8',
-                    'label': {
-                        'text': _('UK postcode')
-                    },
-                    'name': 'request-postcode',
-                    'error': {
-                        'text': _('Enter a valid UK postcode')
-                    },
-                })
-            }}
-        {% else %}
-            {{
-                onsInput({
-                    'id': 'postcode',
-                    'type': 'postcode',
-                    'autocomplete': 'postcode',
-                    'classes': 'input--w-8',
-                    'label': {
-                        'text': _('UK postcode')
-                    },
-                    'name': 'request-postcode'
-                })
-            }}
-        {% endif %}
-
-    {% endcall %}
-
-    {{
-        onsButton({
-            'text': _('Continue'),
-            'classes': 'u-mt-l'
-        })
-    }}
+    {% include 'partials/enter-address.html' with context %}
 
 {% endblock %}
